+++ include/memory/page_alloc.h
@@ -4,13 +4,18 @@
 #include <memory/frame.h>
 #include <memory/paging.h>
 
-/* Allocate a single page in the linear address zone. */
+/*
+ * Allocates a page of virtual memory which is mapped linearly
+ * to physical memory.
+ *
+ * Linear zone guarantees that a page can be mapped to a single frame,
+ * and a frame can be mapped to a single page. So checking that the
+ * zone has a free page, tells us that it must have a free frame.
+ * This is not true of other zones (e.g. if virtual mem >> phys mem)
+ *
+ */
 pg_idx_t alloc_linear(void);
 pg_idx_t free_linear(pg_idx_t pg_to_free);
-
-pg_idx_t alloc_pg_table(pg_idx_t pg_idx);
-
-va_t alloc_adhoc(void);
 
 frame_t *pg_linear_descriptor(va_t pg_va);
